unix port: rename archtecture defines (linux -> LWIP_UNIX_LINUX)
the fine differences between unix implementations used to be covered by
preprocessor macros "linux", "openbsd" and "cygwin". given that gcc
under linux pre-defines a legacy "linux" macro, this can lead to
confusion -- the default setting of -Dopenbsd is often ignored due to
the sequence ifdefs often are used in (if defined linux, else if defined
openbsd), but that is a time bomb.

putting gcc in standard mode using --std=c99 or similar would remove the
macro, but still make the flags hard to spot; besides, the rest of the
code does not work with standard C out of the box.

as a solution, all macros are prefixed with LWIP_UNIX_ followed by an
uppercased variant of the operating system name. in the course of this
change, the defaults are set to LINUX in all make files that use it --
the default was sometimes openbsd, sometimes "generic" before, but
covered by the linux-define of gcc on linux systems.