unix port: tapif: removed rx packet type filtering
It seems to come from a legacy way of pushing packets to the stack, calling
for eg. ethernet_input() or ip_input() or pppoe_input() depending on the
incoming packet type. That's not necessary anymore and a IP stack *must*
be able to deal with any received packet type and whatever level of packet
corruption we might receive.