sio: avoid implicit type cast in sio_read/sio_write
On POSIX systems the result of read() and write() can be
negative to mark an error. The current implementation of
sio_read() and sio_write() will have an implicit type
cast from signed integer to unsigned integer.

Every time, when a error has happened, the upper layer sees
a much too large number of received or sendet bytes. That
leads to a potential buffer overrun or misinterpretation of
what has happened.

Signed-off-by: Stephan Linz <linz@li-pro.net>