sio: add support for SLIP test interface
A SLIP interface should connected to a real serial line. For test
purposes it becomes important to use a PTY that will attached to
a SLIP interface on host side.

The SIO driver has been extendet to a device number (devnum=4) that
will spawn 'slattach' with the proper options similar to pppd.

Signed-off-by: Stephan Linz <linz@li-pro.net>